Moved functions from main.cpp to creature_states.c
