+++ src/creature_states.h
@@ -394,6 +394,10 @@ TbBool creature_will_attack_creature_incl_til_death(const struct Thing *tng1, co
 TbBool creature_is_kept_in_custody_by_enemy_or_dying(const struct Thing *thing);
 
 TbBool creature_state_cannot_be_blocked(const struct Thing *thing);
+
+TbBool setup_move_off_lava(struct Thing* thing);
+TbBool setup_move_out_of_cave_in(struct Thing* thing);
+
 /******************************************************************************/
 #ifdef __cplusplus
 }
